+++ base/organise.py
@@ -63,6 +63,8 @@ class Organise:
         self.add_publisher:str|None = None
         self.clean:list|None = None
         self.prepare:list|None = None
+        self.ba_username:str|None = None
+        self.ba_password:str|None = None
         self.quiet:bool = False
 
         # Set up list of allowed arguments
@@ -184,6 +186,18 @@ class Organise:
             help = 'Prepare cto output for this NFDI4Culture feed and catalog ID'
         )
         available_args.add_argument(
+            '-bu', '--ba_username',
+            default = None,
+            type = str,
+            help = 'Basic Auth username for requests'
+        )
+        available_args.add_argument(
+            '-bp', '--ba_password',
+            default = None,
+            type = str,
+            help = 'Basic Auth password for requests'
+        )
+        available_args.add_argument(
             '-q', '--quiet',
             default = False,
             action = 'store_true',
@@ -209,6 +223,8 @@ class Organise:
         self.add_publisher = args.add_publisher
         self.clean = args.clean
         self.prepare = args.prepare
+        self.ba_username = args.ba_username
+        self.ba_password = args.ba_password
         self.quiet = args.quiet
 
         # Check location based on feed parameter
@@ -217,7 +233,7 @@ class Organise:
                 raise ValueError('Hydra Scraper called with a malformed folder location.')
         elif self.feed in ['beacon', 'cmif', 'csv', 'schema', 'schema-list']:
             if not url(self.location) and not isfile(self.location):
-                raise ValueError('Hydra Scraper called with a malformed ZIP file location.')
+                raise ValueError('Hydra Scraper called with a malformed entry-point location.')
         elif self.feed in ['oaipmh']:
             if not url(self.location):
                 raise ValueError('Hydra Scraper called with a malformed API location.')
@@ -226,6 +242,12 @@ class Organise:
         if not self.elements:
             if 'beacon' in self.output or 'csv' in self.output or 'cto' in self.output:
                 raise ValueError('Hydra Scraper called with extraction routine but no element markup.')
+
+        # Check Basic Auth data
+        if self.ba_username != None and self.ba_password == None:
+            raise ValueError('Hydra Scraper called with Basic Auth username but no password.')
+        elif self.ba_username == None and self.ba_password != None:
+            raise ValueError('Hydra Scraper called with Basic Auth password but no username.')
 
         # Check further URIs
         for uri in [self.add_feed, self.add_catalog, self.add_publisher]: