+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 - Remove local file paths from source-file path
 - Ignore LIDO files that have no LIDO content
-- Fix authentication issues when looking for a `robots.txt`
 - Various fixes to deal with faulty URIs
+- Fix authentication issues when looking for a `robots.txt`
+- Support Basic Auth via `--ba_username` and `--ba_password`
 
 ## 0.9.3